la demo virke for pr også
