+++ src/App.tsx
@@ -6,6 +6,8 @@ import { useAppContext } from './AppContext.tsx'
 import useFetch from './hooks/useFetch.ts'
 import { createPamelding } from './api/api.ts'
 import { Alert, Heading, Loader } from '@navikt/ds-react'
+import { isEnvLocalDemoOrPr } from './utils/environment-utils.ts'
+import DemoBanner from './components/demo-banner/DemoBanner.tsx'
 
 dayjs.locale('nb')
 
@@ -27,17 +29,21 @@ const App = () => {
 
   if (error || !nyPamelding) {
     return (
-      <Alert variant="error">
-        <Heading spacing size="small" level="3">
-          Vi beklager, men noe gikk galt
-        </Heading>
-        {error}
-      </Alert>
+      <>
+        {isEnvLocalDemoOrPr && <DemoBanner hasError />}
+        <Alert variant="error">
+          <Heading spacing size="small" level="3">
+            Vi beklager, men noe gikk galt
+          </Heading>
+          {error}
+        </Alert>
+      </>
     )
   }
 
   return (
     <PameldingContextProvider initialPamelding={nyPamelding}>
+      {isEnvLocalDemoOrPr && <DemoBanner />}
       <DeltakerGuard />
     </PameldingContextProvider>
   )
