+++ src/components/demo-banner/DemoBanner.tsx
@@ -1,7 +1,13 @@
 import { Alert, BodyLong, BodyShort } from '@navikt/ds-react'
 import DemoDeltakerInstillinger from './DemoDeltakerInstillinger'
+import { useMock } from '../../utils/environment-utils'
+import DemoStatusInstillinger from './DemoStatusInstillinger'
 
-const DemoBanner = () => {
+interface Props {
+  hasError?: boolean
+}
+
+const DemoBanner = ({ hasError }: Props) => {
   return (
     <Alert variant="warning" className="mb-4" size="small">
       <BodyShort weight="semibold" size="small">
@@ -12,7 +18,9 @@ const DemoBanner = () => {
         Demoen inneholder ikke ekte data og kan til tider være ustabil.
       </BodyLong>
 
-      <DemoDeltakerInstillinger />
+      {hasError && <DemoDeltakerInstillinger />}
+      {!hasError && useMock && <DemoStatusInstillinger />}
+      {!hasError && !useMock && <DemoDeltakerInstillinger />}
     </Alert>
   )
 }
