+++ src/guards/DeltakerGuard.tsx
@@ -3,8 +3,6 @@ import { OpprettPameldingPage } from '../pages/OpprettPameldingPage.tsx'
 import { RedigerPameldingPage } from '../pages/RedigerPameldingPage.tsx'
 import { TiltakPage } from '../pages/TiltakPage.tsx'
 import { usePameldingCOntext } from '../components/tiltak/PameldingContext.tsx'
-import DemoBanner from '../components/demo-banner/DemoBanner.tsx'
-import { isEnvLocalDemoOrPr } from '../utils/environment-utils.ts'
 
 export const DeltakerGuard = () => {
   const { pamelding } = usePameldingCOntext()
@@ -19,10 +17,5 @@ export const DeltakerGuard = () => {
     pageToLoad = <TiltakPage />
   }
 
-  return (
-    <>
-      {isEnvLocalDemoOrPr && <DemoBanner />}
-      {pageToLoad}
-    </>
-  )
+  return pageToLoad
 }
